Architecture Portofolio
I am a fresh graduate of Bachelor of Architecture at Muhammadiyah University, Surakarta. During college, I was active in participating in organizations and campus event committees. Apart from that, I often attend seminars related to architecture. During college I learned many things, especially in communication. This is my strength, so I can communicate well with new people individually and in groups. I am capable of designing buildings and landscapes using various architectural approaches. I like to explore interesting things, especially in the field of architecture. This portfolio was created with the aim of displaying several works that have been done while studying architecture from 2019 to 2024. This portfolio was created as a forum to present the results of designs created during the architecture study period from 2019 to 2024.